Adding SlamPost Class and related functionatliy
SlamPost is a class with basic added functionality that is useful for
creating classes for custom post types.

The Model Class now extend the SlamPost class, and gets easy access to
meta values.

The Model Class also get a “block()” view.

We added a shortcode “get_models” which displays all models as a block
list.